fixing modules and view engine
