+++ package.json
@@ -7,13 +7,14 @@
   "scripts": {
     "prepare": "husky install",
     "precommit": "run-p lint build",
-    "clean": "rimraf src/**/*.js test/**/*.js db.json build",
+    "clean": "rimraf src/**/*.js test/**/*.js build",
+    "clean:db": "rimraf db.json",
     "lint": "eslint . --ext .ts,.tsx",
     "build": "tsc --project .",
     "postinstall": "ts-node postinstall.ts",
     "prestart": "run-p clean prestart:wait",
     "prestart:wait": "ts-node prestart.ts",
-    "start": "tsnd --watch=.env --inspect=0.0.0.0:9229 --respawn --transpile-only -- src/server.ts",
+    "start": "tsnd --watch=.env,**/*.squirrelly --inspect=0.0.0.0:9229 --respawn --transpile-only -- src/server.ts",
     "test:e2e": "run-s test:e2e:install test:e2e:start",
     "test:e2e:install": "playwright install",
     "test:e2e:start": "playwright test",
@@ -24,7 +25,7 @@
     "atlassian-jwt": "^2.0.2",
     "express": "^4.18.2",
     "lodash": "^4.17.21",
-    "mustache-express": "^1.3.2",
+    "squirrelly": "^8.0.8",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
