--- src/views/config.mst
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/aui/9.5.1/aui/aui-prototyping.min.css" />
-    <link rel="stylesheet" href="/public/css/main.css">
-    <title>{{index}}</title>
-</head>
-<body>
-<br/>
-<br/>
-<button class="aui-button" id="homePage">Home</button>
-<br/>
-<br/>
-<h3>Atlassian Connect Config:</h3>
-<hr>
-<pre class="config-container">{{config}}</pre>
-<br/>
-<br/>
-</body>
-<!-- This script is needed for all the views in the connect app -->
-<script src="https://connect-cdn.atl-paas.net/all.js" async></script>
-<script>
-    /**
-     *  Page navigation is handled by AP.navigator
-     *  source: https://developer.atlassian.com/cloud/jira/software/jsapi/navigator/
-     */
-    document.getElementById('homePage').addEventListener('click', () => {
-        AP.navigator.go('addonmodule', { moduleKey: 'acn-introduction' });
-    });
-</script>
-</html>
