--- src/views/index.mst
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/aui/9.5.1/aui/aui-prototyping.min.css" />
-    <link rel="stylesheet" href="/public/css/main.css">
-    <title>{{index}}</title>
-</head>
-<body>
-<br/>
-<br/>
-<h3>Node app:</h3>
-<hr>
-<br>
-<div id="indexPage">
-    {{body}}
-</div>
-<br>
-<hr>
-<br>
-<div>
-    Check out the links:
-    <ul>
-        <li>
-            <a class="aui-button aui-button-link" id="configPage">Atlassian Connect App config</a>
-        </li>
-        <li>
-            <a class="aui-button aui-button-link" id="webhookLogsPage">Jira Webhook Logs</a>
-        </li>
-    </ul>
-</div>
-<br/>
-<br/>
-</body>
-<!-- This script is needed for all the views in the connect app -->
-<script src="https://connect-cdn.atl-paas.net/all.js" async></script>
-<script>
-    /**
-     *  Page navigation is handled by AP.navigator
-     *  source: https://developer.atlassian.com/cloud/jira/software/jsapi/navigator/
-     */
-    document.getElementById('configPage').addEventListener('click', () => {
-        AP.navigator.go('addonmodule', { moduleKey: 'acn-config' });
-    });
-    document.getElementById('webhookLogsPage').addEventListener('click', () => {
-        AP.navigator.go('addonmodule', { moduleKey: 'acn-logs-webhooks' });
-    });
-</script>
-</html>
