+++ src/views/layout.squirrelly
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>{{ it.title }}</title>
+    <meta charset="UTF-8">
+    <link rel="stylesheet" href="https://unpkg.com/@atlaskit/css-reset/dist/bundle.css" media="all">
+    <!--<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/aui/9.5.1/aui/aui-prototyping.min.css" />-->
+    <!-- This script is needed for all the views in the connect app -->
+    <script src="https://connect-cdn.atl-paas.net/all.js" data-options="sizeToParent:true" async></script>
+
+    <style>
+        body {
+            background: #fff;
+        }
+
+        .ac-content {
+            padding: 2em;
+        }
+    </style>
+</head>
+<body>
+<section class="ac-content">
+{{* it.content }}
+</section>
+<script>
+    /**
+     *  Page navigation is handled by AP.navigator
+     *  source: https://developer.atlassian.com/cloud/jira/software/jsapi/navigator/
+     */
+    document.querySelectorAll('[data-connect-module-key]')
+            .forEach(el =>
+                    el.addEventListener('click', () =>
+                            AP.navigator.go('addonmodule', { moduleKey: el.getAttribute('data-connect-module-key') }))
+            );
+</script>
+</body>
+</html>
