+++ static/css/main.css
@@ -1,8 +1,5 @@
-body {
-    background: #fff;
-    padding: 5% 10%;
-}
 
+/*
 .flexed {
     display: flex;
     justify-content: space-between;
@@ -21,4 +18,4 @@ body {
     color: #fff;
     border-radius: 4px;
     padding: 6px 12px;
-}
+}*/
